Add options for filtering by time
Adds two CLI options for filtering the merged logs to a given range of
time. The first option `-stop=<timestamp>` takes a log message timestamp
that defines the end of time window of interest. The second option
`-duration=<seconds>` take an integer number of seconds. Only messages
in the range `stop - duration` to `stop` are printed.