Add custom scan function for log entries
This commit adds a custom scan function for iterating over the log
messages contained in a file. Previously, this was done by iterating
over the lines in a file and associating lines not containing a
timestamp with the last known timestamp. The custom scan function looks
for the occurences of a newline followed by `[` to separate groups of
lines into single log messages.